+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,26 @@
+> PR message template 및 체크 리스트 입니다.
+> 아래 사항들을 전부 작성/체크 하시고 PR 해주세요!
+
+## 수정 사항 간략한 한줄 요약
+
+(여기에 수정 사항에 대한 간략한 한줄 요약/제목 작성해 주세요.)
+
+- 로그인 버튼 활성화 기능 구현(예시)
+
+## 수정 사항들 자세한 내용
+
+(여기에 수정 사항에 대한 자세한 내용을 작성해 주세요.)
+
+- id '@' 포함, pw 5 글자 이상 입력 시 버튼 활성화(예시)
+
+## 기타 질문 및 특이 사항
+
+(궁금한 사항들, 하면서 느낀 점들 공유해주실 것이 있으면 자유롭게 작성해 주세요.)
+
+- 함수 로직이 너무 긴 것 같습니다. 좀 더 효율적인 방법이 없을지 궁금합니다.(예시)
+
+## 체크 리스트 (아래 사항들이 전부 체크되어야만 merge가 됩니다!)
+
+- [] 필요한 test들을 완료하였고 기능이 제대로 실행되는지 확인 하였습니다.
+- [] 제가 의도한 파일들과 수정 사항들만 커밋이 된 것을 확인 하였습니다.
+- [] 본 수정 사항들을 팀원들과 사전에 상의하였고 팀원들 모두 해당 PR에 대하여 알고 있습니다.
