+++ .gitignore
@@ -17,7 +17,6 @@
 .env.development.local
 .env.test.local
 .env.production.local
-.github
 
 npm-debug.log*
 yarn-debug.log*
