+++ package.json
@@ -12,12 +12,14 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.19.2",
+    "file-saver": "^2.0.2",
     "google-maps-react": "^2.0.6",
     "i": "^0.3.6",
     "npm": "^6.14.8",
     "react": "^16.13.1",
     "react-datepicker": "^3.1.3",
     "react-dom": "^16.13.1",
+    "react-export-excel": "^0.5.3",
     "react-google-login": "^5.1.21",
     "react-google-maps": "^9.4.5",
     "react-google-recaptcha": "^2.1.0",
@@ -30,7 +32,8 @@
     "redux-actions": "^2.6.5",
     "redux-devtools-extension": "^2.13.8",
     "slick-carousel": "^1.8.1",
-    "styled-components": "^5.1.1"
+    "styled-components": "^5.1.1",
+    "xlsx": "^0.16.6"
   },
   "scripts": {
     "start": "react-scripts start",
