+++ src/Pages/Components/CRSP/CodeLookupText.js
@@ -1,0 +1,33 @@
+import React from "react";
+
+export const FirstParagraph = (
+  <>
+    <p>
+      The Company Code Lookup Tool is designed to find all identifiers
+      associated with a specific company. This tool can retrieve identifiers for
+      multiple companies, then add the codes directly to your query. You can
+      also use this tool to save and download codes for future queries.
+      Downloaded codes are saved in the text (.txt) file format.
+    </p>
+  </>
+);
+
+export const SecondParagraph = (
+  <>
+    <ol>
+      <li>
+        Start by entering a company name (or part of the name) into the search
+        box above.
+      </li>
+      <li>
+        After the results are displayed, make your selections and chose an
+        identifier.
+      </li>
+      <li>Next, refine your list of codes or add more.</li>
+      <li>
+        Last, choose whether to insert your selections into your web query or to
+        download them as a text file for later use.
+      </li>
+    </ol>
+  </>
+);
