+++ src/Styles/ModalStyle.js
@@ -1,0 +1,22 @@
+import { css } from "styled-components";
+
+export const ModalStyle = css`
+  position: fixed;
+  top: 0;
+  right: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.3);
+  z-index: 10001;
+  opacity: 1;
+  animation: opacity 0.15s linear;
+
+  @keyframes opacity {
+    from {
+      opacity: 0;
+    }
+    to {
+      opacity: 1;
+    }
+  }
+`;
